Lesson 74

+++ app/Http/Controllers/Admin/RolesController.php
@@ -17,6 +17,8 @@ class RolesController extends Controller
      */
     public function index()
     {
+      $this->authorize('view', new Role);
+
       return view('admin.roles.index', [
         'roles' => Role::all()
       ]);
@@ -29,8 +31,10 @@ class RolesController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', $role = new Role);
+
         return view('admin.roles.create', [
-          'role' => new Role,
+          'role' => $role,
           'permissions' => Permission::pluck('name', 'id')
         ]);
     }
@@ -52,6 +56,7 @@ class RolesController extends Controller
         //   'name.unique' => 'Этот Identificator уже зарегистрирован.',
         //   'display_name.required' => 'Поле Nombre обязательно для заполнения.'
         // ]);
+        $this->authorize('create', new Role);
 
         $role = Role::create($request->validated());
 
@@ -64,17 +69,6 @@ class RolesController extends Controller
     }
 
     /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
-    }
-
-    /**
      * Show the form for editing the specified resource.
      *
      * @param  int  $id
@@ -82,6 +76,8 @@ class RolesController extends Controller
      */
     public function edit(Role $role)
     {
+        $this->authorize('update', $role);
+
         return view('admin.roles.edit', [
           'role' => $role,
           'permissions' => Permission::pluck('name', 'id')
@@ -103,7 +99,8 @@ class RolesController extends Controller
       // ]
       //
       // );
-// dd($request);
+      $this->authorize('update', $role);
+
       $role->update($request->validated());
 
       $role->permissions()->detach();
@@ -125,11 +122,11 @@ class RolesController extends Controller
      */
     public function destroy(Role $role)
     {
-        if ( $role->id === 1 ) {
-          throw new \Illuminate\Auth\Access\AuthorizationException('Эту роль удалить нельзя!');
-        }
+        // if ( $role->id === 1 ) {
+        //   throw new \Illuminate\Auth\Access\AuthorizationException('Эту роль удалить нельзя!');
+        // }
 
-        // $this->authorize('delete', $role);
+        $this->authorize('delete', $role);
 
         $role->delete();
